# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "API documentation for Postaway, a Social Media APP.",
    "title": "Postaway-II API",
    "contact": {
      "name": "Sachin Bhawar",
      "email": "sachinbhavar@gmail.com",
      "url": "https://www.linkedin.com/in/sachin-bhawar/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3600"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "JWT token to be included in the 'Authorization' header in the format 'Bearer token'."
      }
    }
  },
  "paths": {
...
    "/api/posts": {
      "post": {
        "tags": ["Posts"],
        "summary": "Create a new post with image upload functionality.",
        "description": "Create a new post with image upload functionality.",
        "security": [
          {
            "JWT": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "image": {
                    "type": "file",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Post created successfully." },
          "400": { "description": "Caption is required / Image is required" },
          "500": { "description": "Something went wrong while working with database" }
        }
      },
      "get": {
        "tags": ["Posts"],
        "summary": "Retrieve posts for logged-in user.",
        "description": "Retrieve all posts associated with the authenticated user.",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": { "description": "All Posts of Logged-in user retrieved successfully" },
          "404": { "description": "No post found." },
          "500": { "description": "Something went wrong while working with database" }
        }
      }
    },
...
  }
}